Split `other` into `diagnostics` and `connection`
This separation makes more sense to me and allows me to require a feature for running `diagnostics` test.
